requirements.txt
pycryptodome==3.8.2
requests==2.22.0
simplejson==3.16.0
